chore: rename `BenefitCtaLight` into `BenefitCta`
